+++ react/packages/@porsche/ui-kit-react-docs/src/examples/molecules/Input/Types/InputExampleTypesNoPlaceholder.js
@@ -1,21 +1,21 @@
 import React from "react"
-import { Input, Flex } from "@porsche/ui-kit-react"
+import { Input, Grid } from "@porsche/ui-kit-react"
 
 const InputExampleTypesNoPlaceholder = () => {
     return (
-        <Flex gap="grid">
-            <Flex.Item width={4}>
+        <Grid>
+            <Grid.Child size={4}>
                 <Input />
-            </Flex.Item>
+            </Grid.Child>
 
-            <Flex.Item width={4}>
+            <Grid.Child size={4}>
                 <Input basic icon="magnify_glass" />
-            </Flex.Item>
+            </Grid.Child>
 
-            <Flex.Item width={4}>
+            <Grid.Child size={4}>
                 <Input basic unit="km" />
-            </Flex.Item>
-        </Flex>
+            </Grid.Child>
+        </Grid>
     )
 }