+++ react/packages/@porsche/ui-kit-react-docs/src/examples/molecules/Select/EdgeCases/SelectExampleLongValue.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { Flex, Select } from "@porsche/ui-kit-react"
+import { Grid, Select } from "@porsche/ui-kit-react"
 
 const data = [
     { value: "WithoutWhitespace", label: "WithoutWhitespaceWithoutWhitespaceWithoutWhitespaceWithoutWhitespace" },
@@ -11,23 +11,23 @@ const data = [
 
 const SelectExampleLongValue = () => {
     return (
-        <Flex gap="grid">
-            <Flex.Item width={3}>
+        <Grid>
+            <Grid.Child size={3}>
                 <Select options={data} value={data[0].value} />
-            </Flex.Item>
+            </Grid.Child>
 
-            <Flex.Item width={3}>
+            <Grid.Child size={3}>
                 <Select options={data} value={data[1].value} />
-            </Flex.Item>
+            </Grid.Child>
 
-            <Flex.Item width={3}>
+            <Grid.Child size={3}>
                 <Select multi options={data} value={data[0].value} />
-            </Flex.Item>
+            </Grid.Child>
 
-            <Flex.Item width={3}>
+            <Grid.Child size={3}>
                 <Select multi options={data} value={data[1].value} />
-            </Flex.Item>
-        </Flex>
+            </Grid.Child>
+        </Grid>
     )
 }