+++ react/packages/@porsche/ui-kit-react-docs/src/examples/molecules/Select/Single/SelectExampleSingleClearable.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { Flex, Select } from "@porsche/ui-kit-react"
+import { Grid, Select } from "@porsche/ui-kit-react"
 
 const data = [
     { value: "one", label: "One" },
@@ -36,25 +36,25 @@ class SelectExampleSingleError extends React.Component {
 
     render() {
         return (
-            <Flex gap="grid">
-                <Flex.Item width={6}>
+            <Grid>
+                <Grid.Child size={6}>
                     <Select
                         clearable={false}
                         options={data}
                         value={this.state.value1}
                         onChange={this.handleValue1Change}
                     />
-                </Flex.Item>
+                </Grid.Child>
 
-                <Flex.Item width={6}>
+                <Grid.Child size={6}>
                     <Select
                         clearable={false}
                         options={data}
                         value={this.state.value2}
                         onChange={this.handleValue2Change}
                     />
-                </Flex.Item>
-            </Flex>
+                </Grid.Child>
+            </Grid>
         )
     }
 }